setup unittest
